requirements.txt
discord.py
google-generativeai
timer